% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_create_pdf.R
\name{bm_create_pdf}
\alias{bm_create_pdf}
\title{Create a pdf from a list of images}
\usage{
bm_create_pdf(
  pages,
  output = NULL,
  ...,
  paper = c("special", "letter", "a4", "poker"),
  orientation = c("portrait", "landscape"),
  width = 8.5,
  height = 11,
  bg = "white"
)
}
\arguments{
\item{pages}{A list of images (each representing a page).}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{orientation}{Either "portrait" or "landscape".  Ignored if \code{paper = "special"}.}

\item{width, height}{Paper size in inches if \code{paper = "special"}.}

\item{bg}{\code{output} pdf background color.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a blank pdf file.
}
\description{
\code{bm_create_pdf()} creates a pdf from a list of images (each representing a page).
}
\examples{
f1 <- pdf_create_blank(paper = "a4", orientation = "landscape")
f2 <- pdf_create_blank(length = 4L)
unlink(f1)
unlink(f2)
}
