% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_crop_layout.R
\name{bm_crop_layout}
\alias{bm_crop_layout}
\title{Crop out a component from a print-and-play layout}
\usage{
bm_crop_layout(
  page,
  ...,
  layout = "button_shy_cards",
  row = 1L,
  col = 1L,
  bleed = FALSE
)
}
\arguments{
\item{page}{A \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object representing a print-and-play layout page.
Often the output from \code{\link[=pdf_render_bm_pixmap]{pdf_render_bm_pixmap()}}.}

\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{row, col}{The \code{row} and \code{col} of the component in the layout (integers).}

\item{bleed}{Include the bleed (if available).}
}
\value{
A \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object.
}
\description{
\code{bm_crop_layout()} crops out a print-and-play component from a layout.
}
\examples{
\dontrun{# User not expected to have this PDF file
if (requireNamespace("bittermelon", quietly = TRUE)) {
  input <- "A Nice Cuppa - PNP.pdf"
  page <- pdf_render_bm_pixmap(input, page = 4L, dpi = 75)
  card <- bm_crop_layout(page, layout = "button_shy_cards", row = 1L, col = 1L)
  grid::grid.raster(card)
}
}
}
