% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_rm_crosshairs.R
\name{bm_rm_crosshairs}
\alias{bm_rm_crosshairs}
\title{Remove crosshairs from a raster object}
\usage{
bm_rm_crosshairs(x, ..., layout = "poker_3x2_bleed")
}
\arguments{
\item{x}{A raster object coercible to a \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object.}

\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}
}
\value{
A \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object.
}
\description{
\code{bm_rm_crosshairs()} removes crosshairs from a raster object.
}
\details{
\itemize{
\item In order to work the PnP layout needs a solid color bleed.
\item The default layout supports \href{https://greggjewell.itch.io/galdors-grip}{Galdor's Grip} (PnP letter size v1).
}
}
\examples{
if (requireNamespace("bittermelon", quietly = TRUE) &&
    requireNamespace("piecepackr", quietly = TRUE) &&
    utils::packageVersion("piecepackr") >= "1.14.0-5") {
  f1 <- pdf_create_blank(length = 1L, width = 11, height = 8.5)
  f2 <- pdf_add_crosshairs(f1, pages = "all",
                           layout = "poker_3x2_bleed", dpi = 75)
  page <- pdf_render_bm_pixmap(f2, page = 1L, dpi = 75)
  page2 <- bm_rm_crosshairs(page, layout = "poker_3x2_bleed")
  unlink(f1)
  unlink(f2)
}
}
\seealso{
\code{\link[=pdf_rm_crosshairs]{pdf_rm_crosshairs()}} to remove crosshairs from a pdf.
}
