% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_split_layout.R
\name{bm_split_layout}
\alias{bm_split_layout}
\title{Extract components from a print-and-play layout}
\usage{
bm_split_layout(page, ..., layout = "button_shy_cards", bleed = FALSE)
}
\arguments{
\item{page}{A \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object representing a print-and-play layout page.
Often the output from \code{\link[=pdf_render_bm_pixmap]{pdf_render_bm_pixmap()}}.}

\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{bleed}{Include the bleed (if available).}
}
\value{
A \code{\link[bittermelon:bm_list]{bittermelon::bm_list()}} of \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} objects.
}
\description{
\code{bm_split_layout()} extracts print-and-play components from a layout.
}
\examples{
\dontrun{# User not expected to have this PDF file
if (requireNamespace("bittermelon", quietly = TRUE)) {
  input <- "A Nice Cuppa - PNP.pdf"
  page <- pdf_render_bm_pixmap(input, page = 4L, dpi = 75)
  cards <- bm_split_layout(page, layout = "button_shy_cards")
  grid::grid.raster(cards[[1L]])
}
}
}
\seealso{
\code{\link[=bm_crop_layout]{bm_crop_layout()}} to just crop a single component
}
