% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_cropmarks.R
\name{grid_add_cropmarks}
\alias{grid_add_cropmarks}
\title{Draw crop marks around components}
\usage{
grid_add_cropmarks(..., layout = "poker_3x3", bleed = NULL)
}
\arguments{
\item{...}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{bleed}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.
If \code{NULL} defaults to \code{max(max(layout$bleed), 0.125)}.}
}
\value{
\code{NULL} invisibly.
As a side effect draws crop marks to the active graphics device.
}
\description{
\code{grid_add_cropmarks()} adds crop marks to the edges of components of a print-and-play layout.
}
\details{
\itemize{
\item This function draws in \strong{inches} so make sure your graphics device is "big" enough.
}
}
\examples{
if (requireNamespace("piecepackr", quietly = TRUE) &&
    utils::packageVersion("piecepackr") >= "1.14.0-6") {
  grid::grid.newpage()
  vp <- grid::viewport(width=8.5, height=11, default.units="in",
                       x=0.5, y=0.5, just=c("left", "bottom"))
  grid::pushViewport(vp)
  grid_add_rects(layout = "poker_3x3")
  grid_add_cropmarks(layout = "poker_3x3")
  grid::popViewport()
}
}
\seealso{
\code{\link[=pdf_add_cropmarks]{pdf_add_cropmarks()}}, \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}
}
