% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_crosshairs.R
\name{grid_add_crosshairs}
\alias{grid_add_crosshairs}
\title{Draw crosshairs at component corners}
\usage{
grid_add_crosshairs(
  ...,
  layout = "poker_3x3",
  gp = gpar(),
  ch_width = unit(1/6, "in"),
  ch_grob = NULL
)
}
\arguments{
\item{...}{Passed to \code{\link[piecepackr:grid.crosshair]{piecepackr::grid.crosshair()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{gp, ch_width, ch_grob}{Passed to \code{\link[piecepackr:grid.crosshair]{piecepackr::grid.crosshair()}}.
\code{ch_grob} defaults to \code{\link[piecepackr:grid.crosshair]{piecepackr::squaresCrosshairGrob()}}.}
}
\value{
\code{NULL} invisibly.
As a side effect draws crosshairs to the active graphics device.
}
\description{
\code{grid_add_crosshairs()} draws crosshairs at the corners of components of a print-and-play layout.
}
\details{
\itemize{
\item This function draws in \strong{inches} so make sure your graphics device is "big" enough.
}
}
\examples{
if (requireNamespace("piecepackr", quietly = TRUE) &&
    utils::packageVersion("piecepackr") >= "1.14.0-5") {
  grid::grid.newpage()
  vp <- grid::viewport(width=8.5, height=11, default.units="in",
                       x=0.5, y=0.5, just=c("left", "bottom"))
  grid::pushViewport(vp)
  grid_add_crosshairs(layout = "poker_3x3")
  grid::popViewport()
}
}
