% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_add_layout.R
\name{grid_add_layout}
\alias{grid_add_layout}
\title{Draw images to a layout with grid}
\usage{
grid_add_layout(images, ..., layout = "poker_3x3")
}
\arguments{
\item{images}{A named list of images with all the names in the \code{name} column of the \code{layout} argument.  Currently supports raster objects supported by \code{bittermelon} and grid grob objects.}

\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}
}
\value{
\code{NULL} invisibly.
As a side effect draws images to the active graphics device.
}
\description{
\code{grid_add_layout()} draws images to a layout in grid.
}
\seealso{
\code{\link[piecepackr:pmap_piece]{piecepackr::pmap_piece()}} to draw pieces to a layout represented by a \code{\link[piecepackr:pp_cfg]{piecepackr::pp_cfg()}} object.
}
