% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_lines.R
\name{grid_add_lines}
\alias{grid_add_lines}
\title{Draw lines along component edges}
\usage{
grid_add_lines(..., layout = "poker_3x3", gp = gpar())
}
\arguments{
\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{gp}{Passed to \code{\link[grid:grid.segments]{grid::grid.segments()}}.}
}
\value{
\code{NULL} invisibly.
As a side effect draws rectangles to the active graphics device.
}
\description{
\code{grid_add_lines()} draws lines along the components of a print-and-play layout.
}
\details{
\itemize{
\item This function draws in \strong{inches} so make sure your graphics device is "big" enough.
}
}
\examples{
grid::grid.newpage()
vp <- grid::viewport(width=8.5, height=11, default.units="in",
                     x=0.0, y=0.0, just=c("left", "bottom"))
grid::pushViewport(vp)
grid_add_lines(layout = "poker_3x3",
               gp = grid::gpar(lty = "dashed", col = "grey"))
grid::popViewport()
}
\seealso{
\code{\link[grid:grid.segments]{grid::grid.segments()}}
}
