% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_origami.R
\name{grid_add_origami_bb}
\alias{grid_add_origami_bb}
\alias{grid_add_origami}
\title{Draw origami fold symbols}
\usage{
grid_add_origami_bb()

grid_add_origami(
  ...,
  xc = unit(0.5, "npc"),
  yc = unit(0.5, "npc"),
  width = unit(JACKET_4x6_FRONT_WIDTH, "inches"),
  height = unit(JACKET_4x6_HEIGHT, "inches"),
  depth = unit(JACKET_4x6_SPINE_WIDTH, "inches")
)
}
\arguments{
\item{...}{Ignored.}

\item{xc, yc}{Center of the jacket as \code{\link[grid:unit]{grid::unit()}} objects.}

\item{width}{Front/back panel width as a \code{\link[grid:unit]{grid::unit()}} object.}

\item{height}{Jacket height as a \code{\link[grid:unit]{grid::unit()}} object.}

\item{depth}{Spine width as a \code{\link[grid:unit]{grid::unit()}} object.}
}
\value{
\code{NULL} invisibly. As a side effect draws origami fold symbols
on the active graphics device.
}
\description{
\code{grid_add_origami()} draws origami fold symbols for a box jacket.
\code{grid_add_origami_bb()} draws origami fold symbols optimized for
\href{https://sites.google.com/view/boardgamebarrio/home}{Boardgame Barrio's Small Board Game Jackets}.
}
\examples{
grid::grid.newpage()
pnpmisc:::grid_add_jacket_outer()
grid_add_origami()
}
\seealso{
\code{\link[=pdf_add_origami]{pdf_add_origami()}}
}
