% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_grid}
\alias{layout_grid}
\title{Layout data frame for a grid of identically sized components}
\usage{
layout_grid(
  nrow = 2L,
  ncol = 1L,
  width = 2.5,
  height = 3.5,
  bleed = 0,
  paper = c("letter", "a4"),
  orientation = c("landscape", "portrait"),
  angle = 0,
  name = NULL
)
}
\arguments{
\item{nrow, ncol}{Number of rows and columns of print-and-play components (e.g. cards)}

\item{width, height, bleed}{Width, height, and bleed for each print-and-play component (in inches)}

\item{paper, orientation}{Print-and-play paper size and orientation}

\item{angle}{Vector of angles of the pieces in degrees (counter-clockwise) (assigned left-to-right, top-to-bottom)}

\item{name}{Vector of names of each piece (assigned left-to-right, top-to-bottom)}
}
\value{
A data frame with columns "row", "col", "x", "y", "width", "height", "bleed", "paper", and "orientation".
}
\description{
\code{layout_grid()} calculates a layout data frame
for a grid of identically sized print-and-play components.
}
\examples{
# Button Shy card layout
layout_grid(nrow = 2L, ncol = 3L, height = 3.447, width = 2.469, bleed = 0.125)
}
