% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_name_fn}
\alias{layout_name_fn}
\title{Function to generate layout names}
\usage{
layout_name_fn(prefix = "piece.", from = 1L, width = 1L, flag = "0")
}
\arguments{
\item{prefix}{Prefix for the generated names.}

\item{from}{Starting number for the generated name suffix.}

\item{width, flag}{Passed to \code{\link[=formatC]{formatC()}}.}
}
\value{
A function with arguments \code{nrow} and \code{ncol} that would return a character vector of length \code{nrow * ncol}.
}
\description{
\code{layout_name_fn()} returns a function that takes \code{nrow} and \code{ncol} arguments and generates a character vector of \code{nrow * ncol} names.  Intended for use with \code{\link[=layout_grid]{layout_grid()}}.
}
\examples{
layout_name_fn()(nrow = 3, ncol = 3)
layout_name_fn("back_", from = 97, width = 3L)(nrow = 2, ncol = 3)
}
