% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_preset}
\alias{layout_preset}
\alias{layout_names}
\title{Layout data frame for a named preset}
\usage{
layout_preset(preset = "button_shy_cards", ...)

layout_names()
}
\arguments{
\item{preset}{Preset name.  Must be in \code{layout_names()}.}

\item{...}{Passed to \code{\link[=layout_grid]{layout_grid()}}}
}
\value{
A data frame with columns "row", "col", "x", "y", "angle", "width", "height", "bleed", "paper", "orientation", and "name".
}
\description{
\code{layout_preset()} calculates a layout data frame
for a named preset.
\code{layout_names()} returns the supported layout presets.
}
\details{
Unless otherwise indicated assumes letter-sized paper.  Supports the following presets:

\describe{
\item{button_shy_cards}{\href{https://buttonshygames.com/collections/pnps}{Button Shy Games} PnP cards.
\strong{Caveat:} matches many but not \strong{all} of their PnP files e.g. \code{ROVE-PNP-V2.pdf} better matches the \code{poker_3x2_bleed} layout.}
\item{button_shy_rules}{\href{https://buttonshygames.com/collections/pnps}{Button Shy Games} PnP rule booklet pages (8-page booklet).}
\item{button_shy_rules_2x2}{\href{https://buttonshygames.com/collections/pnps}{Button Shy Games} PnP rule booklet pages (4-page booklet).}
\item{poker_3x2_bleed}{Poker-sized cards (2.5" by 3.5") in 3 columns of 2 cards (landscape) with an 1/8" bleed around each card.  Examples of PnP games using this layout include \href{https://greggjewell.itch.io/galdors-grip}{Galdor's Grip}.}
\item{poker_3x3}{Poker-sized cards (2.5" by 3.5") in 3 columns of 3 cards (portrait) with an zero bleed around each card.  Examples of PnP games using this layout include the original \href{https://boardgamegeek.com/boardgame/199242/mini-rogue-a-roguelike-microgame}{Mini Rogue}.}
\item{poker_4x2}{Poker-sized cards (2.5" by 3.5") in 4 columns of 2 cards (landscape) with zero bleed around each card.  Examples of PnP games using this layout include the \href{https://www.decktet.com/getit.php}{Decktet}.}
\item{4x6_jacket}{A single 4x6 photo storage box jacket as created by \code{\link[=pdf_create_4x6_jacket]{pdf_create_4x6_jacket()}}.  The width covers the full jacket (back + spine + front).}
\item{poker_jacket_1x1}{A single poker deck storage box jacket as created by \code{\link[=pdf_create_poker_jacket]{pdf_create_poker_jacket()}}.  The width covers the full jacket (back + spine + front).}
\item{poker_jacket_1x2}{Two poker deck storage box jackets as created by \code{\link[=pdf_create_poker_jacket]{pdf_create_poker_jacket()}} when passed a list of two items.  Uses portrait paper.  The \code{bleed} column is zero (the gap between jackets is encoded in the y positions).}
}
}
\examples{
layout_preset("button_shy_cards")
layout_names()
}
