% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_crosshairs.R
\name{pdf_add_crosshairs}
\alias{pdf_add_crosshairs}
\title{Add crosshairs to a pdf}
\usage{
pdf_add_crosshairs(
  input,
  output = NULL,
  ...,
  layout = "button_shy_cards",
  gp = gpar(),
  ch_width = unit(1/6, "in"),
  ch_grob = NULL
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[=pdf_add_overlay]{pdf_add_overlay()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{gp, ch_width, ch_grob}{Passed to \code{\link[piecepackr:grid.crosshair]{piecepackr::grid.crosshair()}}.
\code{ch_grob} defaults to \code{\link[piecepackr:squaresCrosshairGrob]{piecepackr::squaresCrosshairGrob()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect adds crosshairs to a pdf.
}
\description{
\code{pdf_add_crosshairs()} adds crosshairs to the corners of components of a print-and-play layout.
}
\details{
\itemize{
\item The default layout supports Button Shy games.
}
}
\examples{
if (pnpmisc:::donttest() &&
    requireNamespace("piecepackr", quietly = TRUE) &&
    utils::packageVersion("piecepackr") >= "1.14.0-5") {
  input <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
  output <- pdf_add_crosshairs(input, pages = "odd",
                               layout = "button_shy_cards")
  unlink(input)
  unlink(output)
}
}
