% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_rects.R
\name{pdf_add_rects}
\alias{pdf_add_rects}
\title{Add (round)rects to a pdf}
\usage{
pdf_add_rects(
  input,
  output = NULL,
  ...,
  layout = "poker_3x3",
  pages = "all",
  dpi = 300,
  r = unit(0, "in"),
  gp = gpar(col = "black", fill = NA, lwd = 1)
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{r, gp}{Passed to \code{\link[grid:grid.roundrect]{grid::grid.roundrect()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates pdf file with added origami symbols.
}
\description{
\code{pdf_add_rects()} adds (round)rects around components of a print-and-play layout.
}
\details{
\itemize{
\item Sometimes if you use the same color as a solid background color
this can be used to effectively "remove" unwanted card outlines.
}
}
\examples{
f1 <- pdf_create_blank(length = 2L, paper = "letter")
f2 <- pdf_add_rects(f1, layout = "poker_3x3", dpi = 75)
# "Remove" unwanted card border lines by covering up with white
f3 <- pdf_add_rects(f2, layout = "poker_3x3", dpi = 75, 
                    gp = grid::gpar(col = "white", fill = NA, lwd = 2))
unlink(f1)
unlink(f2)
unlink(f3)
}
