% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_append_blank.R
\name{pdf_append_blank}
\alias{pdf_append_blank}
\title{Append blank pages to a pdf}
\usage{
pdf_append_blank(input, output = NULL, ..., minimum = 1L, multiples_of = 1L)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{minimum}{Final number of pages should be at least this integer.}

\item{multiples_of}{Final number of pages should be a multiple of this integer.}
}
\description{
\code{pdf_append_blank()} appends blank pages to a pdf.
}
\examples{
f1 <- pdf_create_blank(length = 1L)
f2 <- pdf_append_blank(f1, multiples_of = 4L)
qpdf::pdf_length(f2)

# Clean up
unlink(f1)
unlink(f2)
}
