% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_blank.R
\name{pdf_blank}
\alias{pdf_blank}
\title{Create blank pdf pages}
\usage{
pdf_blank(
  output = NULL,
  ...,
  length = 1L,
  width = 8.5,
  height = 11,
  bg = "white"
)
}
\arguments{
\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{length}{Number of pages to create.}

\item{width, height}{the width and height of the graphics region in
    inches.  The default values are \code{7}.}

\item{bg}{\code{output} pdf background color.  Passed to \code{\link[grDevices:pdf]{grDevices::pdf()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a blank pdf file.
}
\description{
\code{pdf_blank()} creates blank pdf pages.
}
\examples{
f <- pdf_blank(length = 4L)
unlink(f) # clean up
}
