% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_clean.R
\name{pdf_clean}
\alias{pdf_clean}
\title{Copies pdf file while removing temporary pdf files}
\usage{
pdf_clean(input, output = NULL, ...)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect copies \code{input} to \code{output} and removes temporary pdf files.
}
\description{
\code{pdf_clean()} copies \code{input} pdf file to \code{output} and removes temporary pdf files with \code{rm_temp_pdfs(exclude = output)}.
}
\examples{
\dontrun{# May delete user files in `tempdir()`
f1 <- pdf_create_blank()
f2 <- pdf_compress(f1)
f3 <- pdf_subset(f2, pages = 1L)
ls_temp_pdfs()
pdf_clean(f3, "output.pdf")
ls_temp_pdfs()
unlink("output.pdf")
}
}
