% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_create_blank.R
\name{pdf_create_blank}
\alias{pdf_create_blank}
\title{Create blank pdf pages}
\usage{
pdf_create_blank(
  output = NULL,
  ...,
  length = 1L,
  paper = c("special", "letter", "a4"),
  orientation = c("portrait", "landscape"),
  width = 8.5,
  height = 11,
  bg = "white"
)
}
\arguments{
\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{length}{Number of pages to create.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{orientation}{Either "portrait" or "landscape".  Ignored if \code{paper = "special"}.}

\item{width, height}{Paper size in inches if \code{paper = "special"}.}

\item{bg}{\code{output} pdf background color.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a blank pdf file.
}
\description{
\code{pdf_create_blank()} creates blank pdf pages.
}
\examples{
f <- pdf_create_blank(length = 4L)
unlink(f) # clean up
}
