% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_create_wallet.R
\name{pdf_create_wallet}
\alias{pdf_create_wallet}
\title{Create print-and-play card wallet pdf}
\usage{
pdf_create_wallet(
  output = NULL,
  ...,
  front = NULL,
  back = NULL,
  spine = NULL,
  bleed = unit(0, "in"),
  paper = c("letter", "a4")
)
}
\arguments{
\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{front}{Fill color/pattern/gradient or grob for front cover section of wallet.
Will be drawn in a rotated masked viewport with a
width of \code{2 + 2 * bleed} inches
and a height of \code{8 + 2 * bleed} inches.
If \code{NULL} (default) we draw a template.}

\item{back}{Fill color/pattern/gradient or grob for back cover section of wallet.
Will be drawn in a rotated masked viewport with a
width of \code{2 * 2 * bleed} inches
and a height of \code{8 + 2 * bleed} inches.
If \code{NULL} (default) we draw a template.}

\item{spine}{Fill color/pattern/gradient or grob for spine section of wallet.
Will be drawn in a masked viewport with a
width of \code{8 + 2 * bleed} inches
and a height of \code{6 + 2 * bleed} inches.
If \code{NULL} (default) we draw a template.}

\item{bleed}{Bleed zone size to assume:
\itemize{
\item{If \code{bleed} is a \code{\link[grid:unit]{grid::unit()}} simply use it}
\item{If \code{bleed} is numeric then convert via \code{grid::unit(bleed, "in")}}
\item{If \code{bleed} is \code{TRUE} assume 1/8 inch bleed zone size}
\item{If \code{bleed} is \code{FALSE} assume 0 inch bleed zone size}
}}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a print-and-play card wallet pdf file.
}
\description{
\code{pdf_create_wallet()} creates print-and-play card origami wallets pdfs.
}
\details{
To make the wallets from the pdf files:
\enumerate{
\item Print out two-sided flipping on the \strong{long} edge
\item Using the crop marks on the first page (the page showing the "spine")
trim the edges with a cutting tool
(the four dots should now be at the corner of each page).
\item Make the four corner valley folds each labeled "Folds 1"
\item Make the two valley folds each labeled "Folds 2"
\item Flip over the paper and make the two valley folds each labeled "Folds 3"
\item Fold in half (valley fold)
}
}
\examples{
# Template `front`, `back`, and `spine`
if (requireNamespace("piecepackr", quietly = TRUE) &&
    getRversion() >= "4.1") {
  f1 <- pdf_create_wallet()
  unlink(f1)
}

# Fill `front`, `back`, and `spine`
if (requireNamespace("piecepackr", quietly = TRUE) &&
    getRversion() >= "4.1") {
f2 <- pdf_create_wallet(front = "red", back = "blue", spine = "purple")
unlink(f2)
}

# Grob `front`, `back`, and `spine` with bleed
if (requireNamespace("piecepackr", quietly = TRUE) &&
    getRversion() >= "4.1" &&
    require("gridpattern", quietly = TRUE)) {
  pal <- grDevices::palette()
  herringbone <- patternGrob("polygon_tiling", type = "herringbone",
                             fill = pal[2], spacing = 0.1)
  rhombille <- patternGrob("polygon_tiling", type = "rhombille",
                           fill = pal[3], spacing = 0.2)
  pythagorean <- patternGrob("polygon_tiling", type = "pythagorean",
                             fill = pal[4], spacing = 0.1)
  f3 <- pdf_create_wallet(front = herringbone, back = rhombille,
                          spine = pythagorean, bleed = TRUE)
  unlink(f3)
}
}
