% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_gs.R
\name{pdf_gs}
\alias{pdf_gs}
\title{Process the pdf file with ghostscript}
\usage{
pdf_gs(input, output = NULL, ..., args = character(0L))
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{args}{Arguments to pass to ghostscript.
Automatically adds \code{-dBATCH -dNOPAUSE -sDEVICE=pdfwrite -sAutoRotatePages=None -sOutputFile={output}}.}
}
\description{
\code{pdf_gs()} processes the pdf file with ghostscript.
This may prevent issues with other pdf processing functions like \code{\link[pdftools:pdftools]{pdftools::pdf_pagesize()}}.
}
\examples{
if (tools::find_gs_cmd()[[1L]] != "") {
  f1 <- pdf_create_blank()
  f2 <- pdf_gs(f1)

  unlink(f1)
  unlink(f2)
}
}
