% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_width.R
\name{pdf_orientation}
\alias{pdf_orientation}
\title{Tell whether pdf is in portrait or landscape mode}
\usage{
pdf_orientation(input, ...)
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}
}
\value{
A character vector with a length equal to the number of pages in \code{input}.
}
\description{
\code{pdf_orientation()} tells whether a pdf is in portrait
or landscape mode.
}
\examples{
f1 <- pdf_create_blank(width = 8.5, height = 11)
pdf_orientation(f1)

f2 <- pdf_create_blank(width = 11, height = 8.5, length = 2L)
pdf_orientation(f2)

unlink(f1)
unlink(f2)
}
