% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pad.R
\name{pdf_pad}
\alias{pdf_pad}
\title{Pad pdf file (to a larger size)}
\usage{
pdf_pad(
  input,
  output = "output.pdf",
  ...,
  bg = "white",
  dpi = 300,
  paper_size = c("letter", "A4")
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.}

\item{...}{Ignored.}

\item{bg}{\code{output} pdf background color.  Passed to \code{\link[grDevices:pdf]{grDevices::pdf()}}.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{paper_size}{Paper size to extend to.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates padded pdf file.
}
\description{
\code{pdf_pad()} makes a pdf file larger by padding it (i.e. adding space to the outside margins).
The original images are \strong{not} rescaled.
}
\examples{
# Some PnP files' size is the intersection of A4/letter page sizes
# i.e. shorter than A4 and narrower than letter.
# We usually want pad to full A4 or letter page size.
input <- tempfile(fileext = ".pdf")
output_letter <- tempfile(fileext = "_letter.pdf")
output_a4 <- tempfile(fileext = "_a4.pdf")
grDevices::pdf(input, width = 8.3, height = 11, bg = "blue")
grid::grid.text("")
invisible(grDevices::dev.off())

pdf_pad(input, output_letter)
pdf_pad(input, output_a4, paper_size = "A4")
}
