% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pad_length.R
\name{pdf_pad_length}
\alias{pdf_pad_length}
\title{Pad pdf with blank pages}
\usage{
pdf_pad_length(input, output = NULL, ..., minimum = 1L, multiples_of = 1L)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{minimum}{Final number of pages should be at least this integer.}

\item{multiples_of}{Final number of pages should be a multiple of this integer.}
}
\description{
\code{pdf_pad_length()} pads a pdf document with blank pages.
}
\examples{
f1 <- pdf_create_blank(length = 1L)
f2 <- pdf_pad_length(f1, multiples_of = 4L)
qpdf::pdf_length(f2)

# Clean up
unlink(f1)
unlink(f2)
}
