% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pages.R
\name{pdf_pages}
\alias{pdf_pages}
\title{Get integer vector of subset of pdf pages}
\usage{
pdf_pages(input, ..., pages = c("all", "even", "odd", "2-up saddle stitch"))
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}
}
\value{
An integer vector.
}
\description{
\code{pdf_pages()} calculates an integer vector of subset of pdf pages.
}
\examples{
f <- pdf_create_blank(length = 8L)
pdf_pages(f, pages = 1:4)
pdf_pages(f, pages = -(1:4))
pdf_pages(f, pages = "all")
pdf_pages(f, pages = "even")
pdf_pages(f, pages = "odd")

unlink(f) # Clean up
}
