% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_paper.R
\name{pdf_paper}
\alias{pdf_paper}
\title{Infer the paper size of a PDF document}
\usage{
pdf_paper(input)
}
\arguments{
\item{input}{Input pdf filename.}
}
\value{
A character vector.
}
\description{
\code{pdf_paper()} infers the paper size of a PDF document
}
\examples{
f <- pdf_create_blank(width = 8.5, height = 11)
pdf_paper(f)
unlink(f)
}
