% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pdfxup.R
\name{pdf_pdfxup}
\alias{pdf_pdfxup}
\alias{pdf_pdfxup_booklet}
\title{N-up pages of a document with \code{pdfxup}}
\usage{
pdf_pdfxup(
  input,
  output = NULL,
  ...,
  ncol = 2L,
  nrow = 1L,
  byrow = TRUE,
  booklet = FALSE,
  orientation = "landscape",
  paper = getOption("papersize", "letter"),
  margins = unit(5, "pt"),
  intspaces = unit(1, "pt"),
  innermargins = unit(5, "pt"),
  framewidth = ifelse(isFALSE(booklet), unit(0.4, "pt"), unit(0, "pt")),
  default.units = "pt",
  clip = TRUE,
  pages = "all",
  bb_pages = "all"
)

pdf_pdfxup_booklet(
  input,
  output = NULL,
  ...,
  booklet = "se",
  margins = unit(0.375, "in"),
  intspaces = unit(0.1875, "in"),
  innermargins = unit(0, "pt")
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{ncol, nrow}{Number of columns/rows per page.}

\item{byrow}{If \code{TRUE} fill pages row by row. If \code{FALSE} fill pages column by column.}

\item{booklet}{If \code{TRUE} configure for printing as a 2-up booklet.
Can also be \code{"se"} (short-edge mode) or \code{"le"} (long-edge mode) for explicit control over two-sided printing; \code{booklet = TRUE} defaults to "le" mode.}

\item{orientation}{Either \code{"landscape"} or \code{"portrait"}.}

\item{paper}{Paper size (e.g. \code{"letter"} or \code{"a4"}).
The name must be known by the LaTeX \code{geometry} package.}

\item{margins}{Margin of pages of the new document as a \code{\link[grid:unit]{grid::unit()}} object.}

\item{intspaces}{Space between pages as a \code{\link[grid:unit]{grid::unit()}} object.}

\item{innermargins}{Inner margin between frame and page as a \code{\link[grid:unit]{grid::unit()}} object.}

\item{framewidth}{Width of the frame around each page as a \code{\link[grid:unit]{grid::unit()}} object.}

\item{default.units}{Default units when a numeric value is passed for
\code{margins}, \code{intspaces}, \code{innermargins}, or \code{framewidth}.}

\item{clip}{If \code{TRUE} clip pages to the computed bounding box.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
}
The functions \code{\link[=pdf_pages]{pdf_pages()}} and \code{\link[=pdf_subset]{pdf_subset()}} also support the string:\describe{
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
\item{interleave}{Interleave the first and second halves of the pages.  Requires an even number of pages.}
\item{interleave_first}{Repeat the first page \code{n - 1} times and interleave with the remaining \code{n - 1} pages.}
\item{interleave_last}{Repeat the last page \code{n - 1} times and interleave with the first \code{n - 1} pages.}
}}

\item{bb_pages}{Pages to use when computing the bounding box (see \code{pages} argument for supported values).
Defaults to \code{"all"}.  Note negative numbers indicate pages to omit from computation of bounding box.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates an n-up pdf.
}
\description{
\code{pdf_pdfxup()} creates an n-up pdf using the \code{pdfxup} command-line program.
\code{pdf_pdfxup_booklet()} is a convenience wrapper around \code{pdf_pdfxup()} with
opinionated defaults better suited for printing a physical booklet on a home printer.
\code{pdfxup} computes the bounding box of the input to remove margins, keeping text readable.
}
\details{
\code{pdf_pdfxup()} tries to closely follow the \code{pdfxup} defaults.  In contrast \code{pdf_pdfxup_booklet()} makes these alternative defaults:
\itemize{
\item \code{booklet = "se"} which I find easier to reason about than \code{"le"}
\item \code{margins = unit(3/8, "in")} to provide adequate margins for home printers
\item \code{intspaces = unit(3/16, "in")} to provide adequate gutter space for folding
\item \code{innermargins = unit(0, "pt")}
}
}
\examples{
f1 <- system.file("doc", "Sweave.pdf", package = "utils")
if (nzchar(Sys.which("pdfxup")) && file.exists(f1)) {
  f2 <- pdf_pdfxup_booklet(f1)
  unlink(f2)
}
}
\seealso{
\url{https://ctan.org/tex-archive/support/pdfxup} for more information about \code{pdfxup}
}
