% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_rasterize.R
\name{pdf_rasterize}
\alias{pdf_rasterize}
\title{Rasterize a pdf}
\usage{
pdf_rasterize(input, output = NULL, ..., dpi = 300)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a rasterized pdf.
}
\description{
\code{pdf_rasterize()} re-renders a pdf by rasterizing each page.
}
\details{
\itemize{
\item The original pdf document will be rasterized.
}
}
\examples{
f1 <- pdf_create_blank(length = 2L, paper = "letter")
f2 <- pdf_rasterize(f1, dpi = 75)
unlink(f1)
unlink(f2)
}
\seealso{
\code{\link[=pdf_apply]{pdf_apply()}} which this function wraps.
}
