% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_render_bm_pixmap.R
\name{pdf_render_bm_list}
\alias{pdf_render_bm_list}
\title{Render all pages of a pdf into bittermelon pixmap objects}
\usage{
pdf_render_bm_list(input, ..., dpi = 300)
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}
}
\value{
A \code{\link[bittermelon:bm_list]{bittermelon::bm_list()}} of \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} objects.
}
\description{
\code{pdf_render_bm_list()} renders all pages of a pdf into bittermelon pixmap objects.
}
\examples{
if (requireNamespace("bittermelon", quietly = TRUE)) {
  f <- pdf_create_wallet()
  bml <- pdf_render_bm_list(f, dpi = 75)
  grid::grid.raster(bml[[1L]])
  unlink(f)
}
}
\seealso{
\code{\link[=pdf_render_bm_pixmap]{pdf_render_bm_pixmap()}}
}
