% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_render_bm_pixmap.R
\name{pdf_render_bm_pixmap}
\alias{pdf_render_bm_pixmap}
\title{Render a pdf page into a bittermelon pixmap object}
\usage{
pdf_render_bm_pixmap(input, ..., page = 1L, dpi = 300)
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}

\item{page}{Integer of page to render.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}
}
\value{
A \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object.
}
\description{
\code{pdf_render_bm_pixmap()} renders a pdf page into a bittermelon pixmap object.
}
\examples{
if (requireNamespace("bittermelon", quietly = TRUE)) {
  f <- pdf_create_wallet()
  bm <- pdf_render_bm_pixmap(f, page = 1L, dpi = 75)
  grid::grid.raster(bm)
  unlink(f)
}
}
\seealso{
\code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}, \code{\link[=pdf_render_raster]{pdf_render_raster()}}
}
