% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_render_bm_pixmap.R
\name{pdf_render_raster}
\alias{pdf_render_raster}
\title{Render a pdf page into a raster object}
\usage{
pdf_render_raster(input, ..., page = 1L, dpi = 300, native = FALSE)
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}

\item{page}{Integer of page to render.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{native}{If \code{TRUE} return a \code{nativeRaster} object else a \code{raster} object.}
}
\value{
If \code{native = TRUE} returns a \code{nativeRaster} object else a \code{raster} object.
}
\description{
\code{pdf_render_raster()} renders a pdf page into a raster object.
}
\examples{
if (requireNamespace("bittermelon", quietly = TRUE)) {
  f <- pdf_create_wallet()
  r <- pdf_render_raster(f, page = 1L, dpi = 75)
  grid::grid.raster(r)
  unlink(f)
}
}
\seealso{
\code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}, \code{\link[=pdf_render_bm_pixmap]{pdf_render_bm_pixmap()}}
}
