% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_rep.R
\name{pdf_rep}
\alias{pdf_rep}
\title{Repeat pages of a pdf}
\usage{
pdf_rep(
  input,
  output = NULL,
  ...,
  pages = "all",
  times = 1L,
  length.out = NA_integer_,
  each = 1L
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
}
The functions \code{\link[=pdf_pages]{pdf_pages()}} and \code{\link[=pdf_subset]{pdf_subset()}} also support the string:\describe{
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
\item{interleave}{Interleave the first and second halves of the pages.  Requires an even number of pages.}
\item{interleave_first}{Repeat the first page \code{n - 1} times and interleave with the remaining \code{n - 1} pages.}
\item{interleave_last}{Repeat the last page \code{n - 1} times and interleave with the first \code{n - 1} pages.}
}}

\item{times, length.out, each}{Passed to \code{\link[base:rep]{base::rep()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a pdf with repeated pages.
}
\description{
\code{pdf_rep()} repeats pages of a pdf similar to \code{\link[base:rep]{base::rep()}}.
}
\examples{
f1 <- pdf_create_blank(length = 3L, paper = "letter")
f2 <- pdf_rep(f1, times = 2L)
f3 <- pdf_rep(f1, pages = 1:2, each = 2L)
unlink(f1)
unlink(f2)
unlink(f3)
}
\seealso{
\code{\link[=pdf_subset]{pdf_subset()}}
}
