% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_scale.R
\name{pdf_scale}
\alias{pdf_scale}
\title{Scale pdf contents centered within the page}
\usage{
pdf_scale(input, output = NULL, ..., scale)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[=pdf_apply]{pdf_apply()}}.}

\item{scale}{A numeric scalar to scale the content by.
Values less than 1 shrink the content within the page; values greater than 1 enlarge it.
By default the original content is \strong{not} rescaled.
Note this also adjusts the size of the \code{\link[grid:viewport]{grid::viewport()}} the \code{grid_fn} will draw in.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a scaled pdf file.
}
\description{
\code{pdf_scale()} scales the contents of a pdf centered within the page.
}
\examples{
input <- pdf_create_blank(paper = "letter", bg = "blue")
output <- pdf_scale(input, scale = 0.9, dpi = 75)
pdf_width(output)
pdf_height(output)
unlink(input)
unlink(output)
}
