% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_width.R
\name{pdf_width}
\alias{pdf_width}
\alias{pdf_height}
\title{Get dimensions of pdf pages}
\usage{
pdf_width(input, ..., units = "inches", numeric = FALSE)

pdf_height(input, ..., units = "inches", numeric = FALSE)
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}

\item{units}{Units to use.  See \code{\link[grid:unit]{grid::unit()}}.}

\item{numeric}{If \code{TRUE} return numeric else a \code{\link[grid:unit]{grid::unit()}} object.}
}
\value{
If \code{numeric = TRUE} a numeric vector else a \code{\link[grid:unit]{grid::unit()}} object.
}
\description{
\code{\link[=pdf_width]{pdf_width()}} and \code{\link[=pdf_height]{pdf_height()}} get the dimensions of the pdf file pages.
}
\examples{
f <- pdf_create_blank(width = 8.5, height = 11)
pdf_width(f)
pdf_height(f, units = "cm")
pdf_height(f, units = "mm", numeric = TRUE)
unlink(f)
}
