% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pnp_pdf}
\alias{pnp_pdf}
\title{Create a pdf device}
\usage{
pnp_pdf(
  output,
  ...,
  width = 8.5,
  height = 11,
  paper = getOption("papersize", "letter"),
  orientation = c("portrait", "landscape"),
  bg = "white"
)
}
\arguments{
\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{width, height}{Passed to underlying graphics device.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{orientation}{Either "portrait" or "landscape".}

\item{bg}{\code{output} pdf background color.}
}
\description{
\code{pnp_pdf()} creates a pdf device preferring \code{\link[grDevices:cairo]{grDevices::cairo_pdf()}}
when supported but falling back to \code{\link[grDevices:pdf]{grDevices::pdf()}} when not.
}
\keyword{internal}
