% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{pdf_combine}
\alias{pdf_combine}
\alias{pdf_compress}
\alias{pdf_overlay_stamp}
\alias{pdf_rotate_pages}
\alias{pdf_subset}
\title{Wrappers around \code{qpdf} functions}
\usage{
pdf_combine(input, output = NULL, ...)

pdf_compress(input, output = NULL, ...)

pdf_overlay_stamp(input, output = NULL, ...)

pdf_rotate_pages(input, output = NULL, ..., pages = "all")

pdf_subset(input, output = NULL, ..., pages = 1L)
}
\arguments{
\item{input}{For \code{pdf_combine()}, a character vector of input pdf filenames.
For all other functions, a single input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to the underlying qpdf functions.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
}
The functions \code{\link[=pdf_pages]{pdf_pages()}} and \code{\link[=pdf_subset]{pdf_subset()}} also support the string:\describe{
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
\item{interleave}{Interleave the first and second halves of the pages.  Requires an even number of pages.}
\item{interleave_first}{Repeat the first page \code{n - 1} times and interleave with the remaining \code{n - 1} pages.}
\item{interleave_last}{Repeat the last page \code{n - 1} times and interleave with the first \code{n - 1} pages.}
}}
}
\value{
\code{output} filename of new pdf file invisibly.
}
\description{
These functions wrap around the utilities in the \code{qpdf} package
but alters the arguments to match this package's conventions:
the first two arguments must be \code{input} and \code{output} (which by default \code{tempfile(fileext = ".pdf")}),
all other arguments must be named,
and any \code{pages} argument is first processed by \code{\link[=pdf_pages]{pdf_pages()}} (so you can also use strings like "even" or "odd").
}
\examples{
f1 <- pdf_create_blank(width = 6, height = 4)
f2 <- pdf_create_blank(width = 4, height = 6)
f3 <- pdf_combine(c(f1, f2))
f4 <- pdf_overlay_stamp(f3, stamp = f1)
f5 <- pdf_compress(f4)
f6 <- pdf_subset(f5, pages = 1L)
f7 <- pdf_rotate_pages(f6, angle = 90)

unlink(f1)
unlink(f2)
unlink(f3)
unlink(f4)
unlink(f5)
unlink(f6)
unlink(f7)
}
\seealso{
\code{\link[qpdf:qpdf]{qpdf::pdf_combine()}}, \code{\link[qpdf:qpdf]{qpdf::pdf_compress()}},
\code{\link[qpdf:qpdf]{qpdf::pdf_overlay_stamp()}}, \code{\link[qpdf:qpdf]{qpdf::pdf_rotate_pages()}}, \code{\link[qpdf:qpdf]{qpdf::pdf_subset()}}
}
