% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_clean.R
\name{ls_temp_pdfs}
\alias{ls_temp_pdfs}
\alias{rm_temp_pdfs}
\title{List and remove temporary pdfs}
\usage{
ls_temp_pdfs(exclude = character())

rm_temp_pdfs(exclude = character())
}
\arguments{
\item{exclude}{A character vector of filenames to exclude.}
}
\value{
\code{ls_temp_pdfs()} returns a character vector.
\code{rm_temp_pdfs()} returns \code{invisible(NULL)}.
}
\description{
Using the functions in this package can lead to several pdf
files cluttering your \code{\link[=tempdir]{tempdir()}}.
\code{ls_temp_pdfs()} lists them while \code{rm_temp_pdfs()} removes them.
}
\examples{
\dontrun{# May delete user pdf files in `tempdir()`
f1 <- pdf_create_blank()
f2 <- pdf_compress(f1)
f3 <- pdf_subset(f2, pages = 1L)
ls_temp_pdfs()
rm_temp_pdfs()
ls_temp_pdfs()
}
}
