% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmpdf.R
\name{pdf_set_bookmarks}
\alias{pdf_set_bookmarks}
\alias{pdf_set_docinfo}
\alias{pdf_set_xmp}
\title{Wrappers around xmpdf functions}
\usage{
pdf_set_bookmarks(input, output = NULL, ..., bookmarks)

pdf_set_docinfo(input, output = NULL, ..., docinfo)

pdf_set_xmp(input, output = NULL, ..., xmp)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Currently ignored.}

\item{bookmarks}{See \code{\link[xmpdf:bookmarks]{xmpdf::set_bookmarks()}}.}

\item{docinfo}{See \code{\link[xmpdf:edit_docinfo]{xmpdf::set_docinfo()}}.}

\item{xmp}{See \code{\link[xmpdf:edit_xmp]{xmpdf::set_xmp()}}.}
}
\value{
\code{output} filename of new pdf file invisibly.
}
\description{
These functions wrap around the utilities in the xmpdf package
but alters the arguments to match this package's conventions:
the first two arguments must be \code{input} and \code{output},
by default \code{output = tempfile(fileext = ".pdf")}, and all other
arguments must be named.
}
\examples{
f1 <- pdf_create_blank(length = 2L)

if (requireNamespace("xmpdf", quietly = TRUE) &&
    xmpdf::supports_set_bookmarks()) {
  bm <- data.frame(title = c("Page 1", "Page 2"), page = c(1, 2))
  f2 <- pdf_set_bookmarks(f1, bookmarks = bm)
  unlink(f2)
}

if (requireNamespace("xmpdf", quietly = TRUE) &&
    xmpdf::supports_set_docinfo()) {
  di <- xmpdf::docinfo(title = "A Title", author = "The Author")
  f3 <- pdf_set_docinfo(f1, docinfo = di)
  unlink(f3)
}

if (requireNamespace("xmpdf", quietly = TRUE) &&
    xmpdf::supports_set_xmp()) {
  xmp <- xmpdf::xmp(title = "A Title", creator = "The Author")
  f4 <- pdf_set_xmp(f1, xmp = xmp)
  unlink(f4)
}
unlink(f1)
}
\seealso{
\code{\link[xmpdf:bookmarks]{xmpdf::set_bookmarks()}}, \code{\link[xmpdf:edit_docinfo]{xmpdf::set_docinfo()}}, \code{\link[xmpdf:edit_xmp]{xmpdf::set_xmp()}}
}
