% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_extract_bm_pixmap.R
\name{zip_extract_bm_pixmap}
\alias{zip_extract_bm_pixmap}
\title{Extract/render an image from a zipfile into a bittermelon pixmap object}
\usage{
zip_extract_bm_pixmap(zipfile, filename, ...)
}
\arguments{
\item{zipfile}{Path to a zip file}

\item{filename}{Filename within zip file}

\item{...}{If \code{filename} is a pdf file then passed to \code{\link[=pdf_render_bm_pixmap]{pdf_render_bm_pixmap()}}}
}
\value{
A \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object.
}
\description{
\code{zip_extract_bm_pixmap()} extracts/renders an image from a zipfile into a \code{\link[bittermelon:bm_pixmap]{bittermelon::bm_pixmap()}} object.
}
\examples{
\dontrun{
# "https://boardgamegeek.com/filepage/279177/packvelopes-storage-boxes"
zipfile <- "Everdeck_Packvelopes.zip"
bm <- zip_extract_bm_pixmap(zipfile, "Red and Black.pdf")

# https://boardgamegeek.com/filepage/62704/cards-in-png-format
zipfile <- "decktet_cards_PNG.zip"
bm <- zip_extract_bm_pixmap(zipfile, "PNG/a_card_cover3.png")
}
}
