% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_grid}
\alias{layout_grid}
\title{Layout data frame for a grid of identically sized components}
\usage{
layout_grid(
  nrow = 2L,
  ncol = 1L,
  width = 2.5,
  height = 3.5,
  bleed = 0,
  paper = getOption("papersize", "letter"),
  orientation = c("landscape", "portrait"),
  angle = 0,
  name = layout_name_fn(),
  direction = "left-to-right"
)
}
\arguments{
\item{nrow, ncol}{Number of rows and columns of print-and-play components (e.g. cards)}

\item{width, height, bleed}{Width, height, and bleed for each print-and-play component (in inches)}

\item{paper, orientation}{Print-and-play paper size and orientation}

\item{angle}{Vector of angles of the pieces in degrees (counter-clockwise) (assigned \code{direction}, top-to-bottom)}

\item{name}{Either a vector of names of each piece or a function with arguments \code{nrow} and \code{ncol} to be used to generate a vector of names for each piece.  These will be assigned \code{direction}, top-to-bottom.  \code{\link[=layout_name_fn]{layout_name_fn()}} can be used to generate such a function.}

\item{direction}{Should \code{name} and \code{angle} be assigned "left-to-right" or "right-to-left".
Also supports aliases "ltr", "lr", "rtl", and "rl".}
}
\value{
A data frame with columns "row", "col", "x", "y", "angle", "width", "height", "bleed", "paper", "orientation", and "name".
}
\description{
\code{layout_grid()} calculates a layout data frame
for a grid of identically sized print-and-play components.
}
\examples{
# Button Shy card layout
layout_grid(nrow = 2L, ncol = 3L, height = 3.447, width = 2.469, bleed = 0.125)
}
