% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_preset}
\alias{layout_preset}
\alias{layout_names}
\title{Layout data frame for a named preset}
\usage{
layout_preset(name = "button_shy_cards")

layout_names()
}
\arguments{
\item{name}{Preset name.  Must be in \code{layout_names()}.}
}
\value{
A data frame with columns "row", "col", "x", "y", "width", "height", "bleed", "paper", and "orientation".
}
\description{
\code{layout_preset()} calculates a layout data frame
for a named preset.
\code{layout_names()} returns the supported layout presets.
}
\examples{
layout_preset("button_shy_cards")
layout_names()
}
