% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_preset}
\alias{layout_preset}
\alias{layout_names}
\title{Layout data frame for a named preset}
\usage{
layout_preset(name = "button_shy_cards")

layout_names()
}
\arguments{
\item{name}{Preset name.  Must be in \code{layout_names()}.}
}
\value{
A data frame with columns "row", "col", "x", "y", "width", "height", "bleed", "paper", and "orientation".
}
\description{
\code{layout_preset()} calculates a layout data frame
for a named preset.
\code{layout_names()} returns the supported layout presets.
}
\details{
Unless otherwise indicated assumes letter-sized paper.  Supports the following presets:

\describe{
\item{button_shy_cards}{\href{https://buttonshygames.com/collections/pnps}{Button Shy Games} PnP cards.
\strong{Caveat:} matches many but not \strong{all} of their PnP files e.g. \code{ROVE-PNP-V2.pdf} better matches the \code{poker_3x2_bleed} layout.}
\item{button_shy_rules}{\href{https://buttonshygames.com/collections/pnps}{Button Shy Games} PnP rule booklet pages (8-page booklet).}
\item{button_shy_rules_2x2}{\href{https://buttonshygames.com/collections/pnps}{Button Shy Games} PnP rule booklet pages (4-page booklet).}
\item{poker_3x2_bleed}{Poker-sized cards (2.5" by 3.5") in 3 columns of 2 cards (landscape) with an 1/8" bleed around each card.  Examples of PnP games using this layout include \href{https://greggjewell.itch.io/galdors-grip}{Galdor's Grip}.}
\item{poker_3x3}{Poker-sized cards (2.5" by 3.5") in 3 columns of 3 cards (portrait) with an zero bleed around each card.  Examples of PnP games using this layout include the original \href{https://boardgamegeek.com/boardgame/199242/mini-rogue-a-roguelike-microgame}{Mini Rogue}.}
\item{poker_4x2}{Poker-sized cards (2.5" by 3.5") in 4 columns of 2 cards (landscape) with zero bleed around each card.  Examples of PnP games using this layout include the \href{https://www.decktet.com/getit.php}{Decktet}.}
}
}
\examples{
layout_preset("button_shy_cards")
layout_names()
}
