% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_cropmarks.R
\name{pdf_add_cropmarks}
\alias{pdf_add_cropmarks}
\title{Add crop marks to a pdf}
\usage{
pdf_add_cropmarks(
  input,
  output = NULL,
  ...,
  layout = "poker_3x3",
  pages = "even",
  rasterize = rasterise,
  dpi = getOption("pnpmisc.dpi", 300),
  paper = NULL,
  bleed = NULL,
  rasterise = NULL
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
}
The functions \code{\link[=pdf_pages]{pdf_pages()}} and \code{\link[=pdf_subset]{pdf_subset()}} also support the string:\describe{
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{rasterize, rasterise}{If \code{TRUE} rasterize the original content using \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.
If \code{FALSE} don't rasterize the original content and throw an error if a requested feature can not yet be implemented without rasterization.
If \code{NULL} only rasterize if a requested feature requires it.
Currently requires rasterization in the following cases:
\enumerate{
\item \code{!is.null(paper)} or \code{scale != 1}
\item \code{any(pages != "all")}
\item \code{!missing(bm_fn)}
}

\code{rasterise} is an alias for \code{rasterize}.}

\item{dpi}{Dots per inch.
Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}} when we rasterize the original content.
Defaults to \code{getOption("pnpmisc.dpi", 300)}.}

\item{paper}{If \code{NULL} (default) the output pdf will be the same size as the input pdf.
Otherwise the output pdf will be created with this paper size
(see \code{\link[=pdf_pad_paper]{pdf_pad_paper()}} for supported values but most commonly "letter" or "a4").}

\item{bleed}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.
If \code{NULL} defaults to \code{max(max(layout$bleed), 0.125)}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect adds crop marks to a pdf.
}
\description{
\code{pdf_add_cropmarks()} adds crop marks to the edges of components of a print-and-play layout.
}
\details{
\itemize{
\item The original pdf document may be rasterized depending on the value of
\code{rasterize}, \code{pages}, and/or \code{paper}.
}
}
\examples{
if (requireNamespace("piecepackr", quietly = TRUE)) {
  input <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
  output <- pdf_add_cropmarks(input, pages = "odd",
                              layout = "button_shy_cards", dpi = 75)
  unlink(input)
  unlink(output)
}
}
\seealso{
\code{\link[=grid_add_cropmarks]{grid_add_cropmarks()}}, \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}
}
