% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_cropmarks.R
\name{pdf_add_cropmarks}
\alias{pdf_add_cropmarks}
\title{Add crop marks to a pdf}
\usage{
pdf_add_cropmarks(
  input,
  output = NULL,
  ...,
  layout = "poker_3x3",
  gp = gpar(),
  cm_width = unit(0.25, "mm"),
  cm_length = unit(0.125, "in")
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[=pdf_add_overlay]{pdf_add_overlay()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{gp, cm_width, cm_length}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect adds crop marks to a pdf.
}
\description{
\code{pdf_add_cropmarks()} adds crop marks to the edges of components of a print-and-play layout.
}
\examples{
if (requireNamespace("piecepackr", quietly = TRUE)) {
  input <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
  output <- pdf_add_cropmarks(input, pages = "odd",
                              layout = "button_shy_cards")
  unlink(input)
  unlink(output)
}
}
\seealso{
\code{\link[=grid_add_cropmarks]{grid_add_cropmarks()}}, \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}
}
