% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_cropmarks.R
\name{pdf_add_cropmarks}
\alias{pdf_add_cropmarks}
\title{Add crop marks to a pdf}
\usage{
pdf_add_cropmarks(
  input,
  output = NULL,
  ...,
  layout = "poker_3x3",
  pages = "even",
  dpi = 300,
  bleed = NULL
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{bleed}{Passed to \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}.
If \code{NULL} defaults to \code{max(max(layout$bleed), 0.125)}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect adds crop marks to a pdf.
}
\description{
\code{pdf_add_cropmarks()} adds crop marks to the edges of components of a print-and-play layout.
}
\details{
\itemize{
\item The original pdf document will be rasterized.
}
}
\examples{
if (requireNamespace("piecepackr", quietly = TRUE)) {
  input <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
  output <- pdf_add_cropmarks(input, pages = "odd",
                              layout = "button_shy_cards", dpi = 75)
  unlink(input)
  unlink(output)
}
}
\seealso{
\code{\link[=grid_add_cropmarks]{grid_add_cropmarks()}}, \code{\link[piecepackr:grid.cropmark]{piecepackr::grid.cropmark()}}
}
