% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_crosshairs.R
\name{pdf_add_crosshairs}
\alias{pdf_add_crosshairs}
\title{Add crosshairs to a pdf}
\usage{
pdf_add_crosshairs(
  input,
  output = NULL,
  ...,
  layout = "button_shy_cards",
  pages = "even",
  dpi = 300
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[piecepackr:grid.crosshair]{piecepackr::grid.crosshair()}}.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect removes from crosshairs from a pdf.
}
\description{
\code{pdf_add_crosshairs()} adds crosshairs to the corners of components of a print-and-play layout.
}
\details{
\itemize{
\item The default layout supports Button Shy games.
\item The original pdf document will be rasterized.
}
}
\examples{
if (requireNamespace("piecepackr", quietly = TRUE) &&
    utils::packageVersion("piecepackr") >= "1.14.0-5") {
  input <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
  output <- pdf_add_crosshairs(input, pages = "odd",
                               layout = "button_shy_cards", dpi = 75)
  unlink(input)
  unlink(output)
}
}
