% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_lines.R
\name{pdf_add_lines}
\alias{pdf_add_lines}
\title{Add lines to a pdf}
\usage{
pdf_add_lines(
  input,
  output = NULL,
  ...,
  layout = "poker_3x3",
  pages = "all",
  dpi = 300,
  gp = gpar()
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored for now.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{gp}{Passed to \code{\link[grid:grid.segments]{grid::grid.segments()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates pdf file with added line segments.
}
\description{
\code{pdf_add_lines()} adds lines along the components of a print-and-play layout.
}
\examples{
f1 <- pdf_create_blank(length = 2L, paper = "letter")
f2 <- pdf_add_lines(f1, layout = "poker_3x3", dpi = 75)
unlink(f1)
unlink(f2)
}
\seealso{
\code{\link[=grid_add_lines]{grid_add_lines()}}, \code{\link[=pdf_add_rects]{pdf_add_rects()}}
}
