% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_origami.R
\name{pdf_add_origami}
\alias{pdf_add_origami}
\title{Add origami symbols to a pdf}
\usage{
pdf_add_origami(
  input,
  output = NULL,
  ...,
  rasterize = rasterise,
  dpi = getOption("pnpmisc.dpi", 300),
  paper = NULL,
  rasterise = NULL
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{rasterize, rasterise}{If \code{TRUE} rasterize the original content using \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.
If \code{FALSE} don't rasterize the original content and throw an error if a requested feature can not yet be implemented without rasterization.
If \code{NULL} only rasterize if a requested feature requires it.
Currently requires rasterization in the following cases:
\enumerate{
\item (\code{!is.null(paper)} or \code{scale != 1}) and ghostscript is unavailable (see \code{\link[=pdf_gs]{pdf_gs()}})
\item \code{any(pages != "all")} and \code{!identical(grid_fn, grid::grid.null)}
\item \code{!identical(bm_fn, identity)}
}

\code{rasterise} is an alias for \code{rasterize}.}

\item{dpi}{Dots per inch.
Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}} when we rasterize the original content.
Defaults to \code{getOption("pnpmisc.dpi", 300)}.}

\item{paper}{If \code{NULL} (default) the output pdf will be the same size as the input pdf.
Otherwise the output pdf will be created with this paper size
(see \code{\link[=pdf_pad_paper]{pdf_pad_paper()}} for supported values but most commonly "letter" or "a4").}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates pdf file with added origami symbols.
}
\description{
\code{pdf_add_origami()} adds origami symbols to the pdf.
Currently only supports adding origami symbols to
\href{https://sites.google.com/view/boardgamebarrio/home}{Boardgame Barrio's Small Board Game Jackets}.
}
\examples{
f1 <- pnpmisc:::pdf_create_mock_sbgj()
f2 <- pdf_add_origami(f1)

unlink(f1)
unlink(f2)
}
