% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_origami.R
\name{pdf_add_origami}
\alias{pdf_add_origami}
\title{Add origami symbols to pdf}
\usage{
pdf_add_origami(input, output = NULL, ..., dpi = 300)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates pdf file with added origami symbols.
}
\description{
\code{pdf_add_origami()} adds origami symbols to the pdf.
Currently only supports adding origami symbols to
\href{https://sites.google.com/view/boardgamebarrio/home}{Boardgame Barrio's Small Board Game Jackets}.
}
\examples{
f1 <- pnpmisc:::pdf_create_mock_sbgj()
f2 <- pdf_add_origami(f1)

unlink(f1)
unlink(f2)
}
