% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_add_overlay.R
\name{pdf_add_overlay}
\alias{pdf_add_overlay}
\title{Overlay custom graphics on a pdf}
\usage{
pdf_add_overlay(
  input,
  output = NULL,
  ...,
  pages = "all",
  dpi = 300,
  grid_fn = grid::grid.null
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{grid_fn}{A zero-argument function called to draw graphics on selected pages.
Defaults to \code{\link[grid:grid.null]{grid::grid.null()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a pdf with overlaid graphics.
}
\description{
\code{pdf_add_overlay()} renders each page of a pdf as a raster image and calls
\code{grid_fn} on selected pages to draw custom graphics on top.
}
\details{
\itemize{
\item The original pdf document will be rasterized.
}
}
\examples{
f1 <- pdf_create_blank(length = 2L, paper = "letter")
f2 <- pdf_add_overlay(f1, pages = "all", dpi = 75,
                      grid_fn = function() {
                        grid::grid.text("DRAFT", gp = grid::gpar(col = "red",
                                        fontsize = 72, alpha = 0.3))
                      })
unlink(f1)
unlink(f2)
}
\seealso{
\code{\link[=pdf_add_cropmarks]{pdf_add_cropmarks()}}, \code{\link[=pdf_add_crosshairs]{pdf_add_crosshairs()}},
\code{\link[=pdf_add_lines]{pdf_add_lines()}}, \code{\link[=pdf_add_rects]{pdf_add_rects()}}
}
