% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_apply.R
\name{pdf_apply}
\alias{pdf_apply}
\title{Modify selected pages on a pdf}
\usage{
pdf_apply(
  input,
  output = NULL,
  ...,
  pages = "all",
  dpi = 300,
  bm_fn = identity,
  grid_fn = grid::grid.null
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{bm_fn}{A one-argument function called to edit the selected pages rasterized.  The input will be a "raster" object and the output must be an object coercible to a "raster" object by \code{\link[grDevices:as.raster]{grDevices::as.raster()}}}

\item{grid_fn}{A zero-argument function called to draw graphics on selected pages.
Defaults to \code{\link[grid:grid.null]{grid::grid.null()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a pdf with overlaid graphics.
}
\description{
\code{pdf_apply()} renders each page of a pdf as a raster image and
on selected pages calls \code{bm_fn()} to modify those raster images
as well as \code{grid_fn} overlay custom graphics on top.
}
\details{
\itemize{
\item The original pdf document will be rasterized.
}
}
\examples{
f1 <- pdf_create_blank(length = 2L, paper = "letter")
if (requireNamespace("bittermelon", quietly = TRUE)) {
  f2 <- pdf_apply(f1, pages = "all", dpi = 75,
                      bm_fn = bittermelon::bm_invert)
  unlink(f2)
}
unlink(f1)
}
\seealso{
\code{\link[=pdf_add_cropmarks]{pdf_add_cropmarks()}}, \code{\link[=pdf_add_crosshairs]{pdf_add_crosshairs()}},
\code{\link[=pdf_add_lines]{pdf_add_lines()}}, \code{\link[=pdf_add_rects]{pdf_add_rects()}}
}
