% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_apply.R
\name{pdf_apply}
\alias{pdf_apply}
\title{Modify selected pages on a pdf}
\usage{
pdf_apply(
  input,
  output = NULL,
  ...,
  pages = "all",
  rasterize = rasterise,
  dpi = getOption("pnpmisc.dpi", 300),
  paper = NULL,
  bg = "white",
  scale = 1,
  bm_fn = identity,
  grid_fn = grid::grid.null,
  rasterise = NULL
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
}
The functions \code{\link[=pdf_pages]{pdf_pages()}} and \code{\link[=pdf_subset]{pdf_subset()}} also support the string:\describe{
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{rasterize, rasterise}{If \code{TRUE} rasterize the original content using \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.
If \code{FALSE} don't rasterize the original content and throw an error if a requested feature can not yet be implemented without rasterization.
If \code{NULL} only rasterize if a requested feature requires it.
Currently requires rasterization in the following cases:
\enumerate{
\item (\code{!is.null(paper)} or \code{scale != 1}) and ghostscript is unavailable (see \code{\link[=pdf_gs]{pdf_gs()}})
\item \code{any(pages != "all")} and \code{!identical(grid_fn, grid::grid.null)}
\item \code{!identical(bm_fn, identity)}
}

\code{rasterise} is an alias for \code{rasterize}.}

\item{dpi}{Dots per inch.
Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}} when we rasterize the original content.
Defaults to \code{getOption("pnpmisc.dpi", 300)}.}

\item{paper}{If \code{NULL} (default) the output pdf will be the same size as the input pdf.
Otherwise the output pdf will be created with this paper size
(see \code{\link[=pdf_pad_paper]{pdf_pad_paper()}} for supported values but most commonly "letter" or "a4").}

\item{bg}{\code{output} pdf background color.}

\item{scale}{A numeric scalar to scale the content by.
Values less than 1 shrink the content within the page; values greater than 1 enlarge it.
By default the original content is \strong{not} rescaled.
Note this also adjusts the size of the \code{\link[grid:viewport]{grid::viewport()}} the \code{grid_fn} will draw in.}

\item{bm_fn}{A one-argument function called to edit the selected pages rasterized.  The input will be a "raster" object and the output must be an object coercible to a "raster" object by \code{\link[grDevices:as.raster]{grDevices::as.raster()}}}

\item{grid_fn}{A zero-argument function called to draw graphics on selected pages.
Defaults to \code{\link[grid:grid.null]{grid::grid.null()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a pdf with overlaid graphics.
}
\description{
\code{pdf_apply()} renders each page of a pdf as a raster image and
on selected \code{pages} additionally calls \code{bm_fn()} to modify those raster images
as well as \code{grid_fn} overlay custom graphics on top.
}
\details{
\itemize{
\item The original pdf document may be rasterized depending on the value of
\code{bm_fn}, \code{pages}, \code{paper}, \code{scale}, and/or \code{rasterize} as well as
whether ghostscript is available (see \code{\link[=pdf_gs]{pdf_gs()}}).
}
}
\examples{
f1 <- pdf_create_blank(length = 2L, paper = "letter", bg = "magenta")
if (requireNamespace("bittermelon", quietly = TRUE)) {
  f2 <- pdf_apply(f1, pages = "even", dpi = 75,
                  bm_fn = bittermelon::bm_invert,
                  grid_fn = function() {
                      grid::grid.text("DRAFT", gp = grid::gpar(col = "red",
                                      fontsize = 72, alpha = 0.3))
                  })
  unlink(f2)
}
unlink(f1)
}
