% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_change_paper.R
\name{pdf_change_paper}
\alias{pdf_change_paper}
\alias{pdf_pad_paper}
\title{Change pdf paper size}
\usage{
pdf_change_paper(
  input,
  output = NULL,
  ...,
  paper = getOption("papersize", "letter")
)

pdf_pad_paper(
  input,
  output = NULL,
  ...,
  paper = getOption("papersize", "letter")
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to \code{\link[=pdf_apply]{pdf_apply()}}.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates a pdf with the new paper size.
}
\description{
\code{pdf_change_paper()} changes the paper size of a pdf, centering the original
contents on the new paper size.
\code{pdf_pad_paper()} is an alias semantically intended for padding an undersized pdf document
to a normal paper size (e.g. padding a pdf smaller than letter to letter size).
}
\details{
By default these function do \strong{not} scale the original contents
(but you can pass on a \code{scale} argument in the \code{...} to \code{\link[=pdf_apply]{pdf_apply()}}).
}
\examples{
# Some PnP files' size is the intersection of A4/letter page sizes
# i.e. shorter than A4 and narrower than letter.
# We usually want pad to full A4 or letter page size.
input <- tempfile(fileext = ".pdf")
grDevices::pdf(input, width = 8.3, height = 11, bg = "blue")
grid::grid.text("")
invisible(grDevices::dev.off())

pdf_width(input)
pdf_height(input)

output <- pdf_change_paper(input, paper = "letter")
pdf_width(output)
pdf_height(output)
unlink(output)

output_a4 <- pdf_pad_paper(input, paper = "a4")
pdf_width(output_a4)
pdf_height(output_a4)
unlink(output_a4)

unlink(input)
}
