% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_create_jacket.R
\name{pdf_create_jacket}
\alias{pdf_create_jacket}
\title{Create printable 4x6 photo box jacket pdf}
\usage{
pdf_create_jacket(
  output = NULL,
  ...,
  front = NULL,
  back = NULL,
  spine = NULL,
  paper = c("letter", "a4")
)
}
\arguments{
\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{front}{Fill color/pattern/gradient or grob for front cover section of wallet.
Will be drawn in a viewport with a
width of 4.139 inches
and a height of 6.141 inches.
If \code{NULL} (default) we draw a template.}

\item{back}{Fill color/pattern/gradient or grob for back cover section of wallet.
Will be drawn in a viewport with a
width of 4.139 inches
and a height of 6.141 inches.
If \code{NULL} (default) we draw a template.}

\item{spine}{Fill color/pattern/gradient or grob for spine section of wallet.
Will be drawn in a rotated viewport with a
width of 6.141 inches
and a height of 1.052 inches.
If \code{NULL} (default) we draw a template.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}
}
\description{
\code{pdf_create_jacket()} creates a printable 4x6 photo box jacket.
}
\details{
To make the 4x6 photo jacket from the pdf file:
\enumerate{
\item Print it out
\item Use the crop marks to trim off the left and right edge
\item Make the two indicated mountain folds on both sides of the spine
\item Trim off the top and bottom edge
\item Insert into the 4x6 photo box
}
}
\examples{
# Template `front`, `back`, and `spine`
f1 <- pdf_create_jacket()
unlink(f1)

# Fill `front`, `back`, and `spine`
f2 <- pdf_create_jacket(front = "red", back = "blue", spine = "purple")
unlink(f2)

# Grob `front`, `back`, and `spine`
if (require("gridpattern", quietly = TRUE)) {
  pal <- grDevices::palette()
  herringbone <- patternGrob("polygon_tiling", type = "herringbone", 
                             fill = pal[2], spacing = 0.1)
  rhombille <- patternGrob("polygon_tiling", type = "rhombille",
                           fill = pal[3], spacing = 0.3)
  pythagorean <- patternGrob("polygon_tiling", type = "pythagorean",
                             fill = pal[4], spacing = 0.1)
  f3 <- pdf_create_jacket(front = herringbone, back = pythagorean,
                          spine = rhombille, bleed = TRUE)
  unlink(f3)
}
}
\seealso{
\code{\link[=pdf_add_origami]{pdf_add_origami()}} to add origami symbols to pre-existing \href{https://sites.google.com/view/boardgamebarrio/home}{Boardgame Barrio's Small Board Game Jackets}.
}
