% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_create_jacket.R,
%   R/pdf_create_jacket_instructions.R
\name{pdf_create_jacket}
\alias{pdf_create_jacket}
\alias{pdf_create_4x6_jacket}
\alias{pdf_create_poker_jacket}
\alias{pdf_create_jacket_instructions}
\title{Create printable storage box jacket pdf}
\usage{
pdf_create_jacket(
  output = NULL,
  ...,
  front = NULL,
  back = NULL,
  spine = NULL,
  inner = NULL,
  paper = c("letter", "a4"),
  orientation = "landscape",
  width = unit(JACKET_4x6_FRONT_WIDTH, "in"),
  height = unit(JACKET_4x6_HEIGHT, "in"),
  depth = unit(JACKET_4x6_SPINE_WIDTH, "in"),
  bg = "transparent"
)

pdf_create_4x6_jacket(
  output = NULL,
  ...,
  front = NULL,
  back = NULL,
  spine = NULL,
  inner = NULL,
  paper = c("letter", "a4")
)

pdf_create_poker_jacket(
  output = NULL,
  ...,
  front = NULL,
  back = NULL,
  spine = NULL,
  inner = NULL,
  paper = c("letter", "a4"),
  depth = unit(JACKET_POKER_SPINE_WIDTH, "in"),
  bg = "transparent"
)

pdf_create_jacket_instructions(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  orientation = "landscape",
  style = marquee::classic_style()
)
}
\arguments{
\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{front}{Fill color/pattern/gradient or grob for front cover section of jacket
Will be drawn in a viewport with a
width of \code{width} and a height of \code{height}.
If \code{NULL} (default) we draw a template.
For \code{pdf_create_poker_jacket()} may also be a list with one or two of these values (in which case will generate two jackets).}

\item{back}{Fill color/pattern/gradient or grob for back cover section of jacket
Will be drawn in a viewport with a
width of \code{width} and a height of \code{height}.
If \code{NULL} (default) we draw a template.
For \code{pdf_create_poker_jacket()} may also be a list with one or two of these values (in which case will generate two jackets).}

\item{spine}{Fill color/pattern/gradient or grob for spine section of jacket
Will be drawn in a rotated viewport with a
width of \code{height} and a height of \code{depth}.
If \code{NULL} (default) we draw a template.
For \code{pdf_create_poker_jacket()} may also be a list with one or two of these values (in which case will generate two jackets).}

\item{inner}{Fill color/pattern/gradient or grob for inner section of jacket.
Will be drawn in a viewport with a width of \code{2 * width + depth} inches
and a height of \code{height} inches.
If \code{NULL} (default) we do not create page for inner section.
For \code{pdf_create_poker_jacket()} may also be a list with one or two of these values (in which case will generate two jackets).}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{orientation}{Either "portrait" or "landscape".  Ignored if \code{paper = "special"}.}

\item{width}{Width of the jacket face as a grid unit.
For a 4x6 photo storage box a good value is 4.139 inches and for a poker deck storage box a good value is 2.6 inches.}

\item{height}{Height of the jacket face as a grid unit.
For a 4x6 photo storage box a good value is 6.141 inches and for a poker deck storage box a good value is 3.6 inches.}

\item{depth}{Width of the spine as a grid unit.
For a 4x6 photo storage box a good value is 1.052 inches and for a poker deck storage box a good value is 0.78 inches.}

\item{bg}{Background fill for the storage jacket.
Passed to \code{\link[=grid.full]{grid.full()}}.}

\item{style}{A style set such as \code{\link[marquee:classic_style]{marquee::classic_style()}} to
be passed to \code{\link[marquee:marquee_grob]{marquee::marquee_grob()}}.}
}
\description{
\code{pdf_create_jacket()} creates a printable storage box jacket.
\code{pdf_create_4x6_jacket()} is a wrapper targeting a 4x6 photo storage box.
\code{pdf_create_poker_jacket()} is an alternative targeting one or two poker deck storage boxes.
\code{pdf_create_jacket_instructions()} creates a printable sheet of
instructions for making such a jacket
(these instructions are repeated in the \code{Details} section below).
}
\details{
To make the storage box jacket from the pdf file:
\enumerate{
\item Print the jacket out (ideally on cardstock).
\itemize{
\item Print "actual size" (100\%).
\item To include the jacket inside (if any) print double-sided.
}
\item Use the crop marks and origami symbols to cut and fold the jacket.
\itemize{
\item Exact procedure depends on your tools and preferred methods.
\item Example procedure #1:
\enumerate{
\item Use the origami mountain fold lines to score
the fold lines (e.g. using a ruler and a penknife).
\item Use the crop marks to cut the edges off.
\item Make two mountain folds using the scored lines along the spine.
}
\item Example procedure #2:
\enumerate{
\item Use the crop marks to cut away the left and right edge.
\item Make a \emph{valley} fold by touching the origami dots symbols
along the top edge together
and then flatten the paper
to make the fold and then unfold flat.
Repeat this procedure with the origami dot symbols along the bottom edge.
\item Cut away the top and bottom edges
(you may use the ends of the
origami mountain fold lines as well as the edges of the jacket
as guide to where to make these cuts).
\item Make two \emph{mountain} folds by reversing the (valley) creases.
}
}
\item Insert the jacket into the storage box.
}
}
\examples{
# Template `front`, `back`, and `spine`
if (requireNamespace("piecepackr", quietlyr= TRUE)) {
  f1 <- pdf_create_4x6_jacket()
  unlink(f1)
}

# Fill `front`, `back`, and `spine`
if (requireNamespace("piecepackr", quietly = TRUE)) {
  f2 <- pdf_create_4x6_jacket(front = "red", back = "blue", spine = "purple")
  unlink(f2)
}

# Grob `front`, `back`, and `spine`
if (requireNamespace("piecepackr", quietly = TRUE) &&
    require("gridpattern", quietly = TRUE)) {
  pal <- grDevices::palette()
  herringbone <- patternGrob("polygon_tiling", type = "herringbone",
                             fill = pal[2], spacing = 0.1)
  rhombille <- patternGrob("polygon_tiling", type = "rhombille",
                           fill = pal[3], spacing = 0.3)
  pythagorean <- patternGrob("polygon_tiling", type = "pythagorean",
                             fill = pal[4], spacing = 0.1)
  f3 <- pdf_create_4x6_jacket(front = herringbone, back = pythagorean,
                          spine = rhombille, bleed = TRUE)
  unlink(f3)
}
}
\seealso{
\code{\link[=pdf_add_origami]{pdf_add_origami()}} to add origami symbols to pre-existing \href{https://sites.google.com/view/boardgamebarrio/home}{Boardgame Barrio's Small Board Game Jackets}.
}
