% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pad_paper.R
\name{pdf_pad_paper}
\alias{pdf_pad_paper}
\title{Pad pdf file (to a larger paper size)}
\usage{
pdf_pad_paper(
  input,
  output = NULL,
  ...,
  bg = "white",
  dpi = 300,
  paper = getOption("papersize", "letter")
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{bg}{\code{output} pdf background color.}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}

\item{paper}{Paper size.  One of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}
}
\value{
\code{output} pdf file name invisibly.
As a side effect creates padded pdf file.
}
\description{
\code{pdf_pad_paper()} makes a pdf file larger by padding it (i.e. adding space to the outside margins).
The original images are \strong{not} rescaled.
}
\examples{
# Some PnP files' size is the intersection of A4/letter page sizes
# i.e. shorter than A4 and narrower than letter.
# We usually want pad to full A4 or letter page size.
input <- tempfile(fileext = ".pdf")
grDevices::pdf(input, width = 8.3, height = 11, bg = "blue")
grid::grid.text("")
invisible(grDevices::dev.off())

pdf_width(input)
pdf_height(input)

output <- pdf_pad_paper(input, dpi = 75)
pdf_width(output)
pdf_height(output)
unlink(output)

output_a4 <- pdf_pad_paper(input, dpi = 75, paper = "a4")
pdf_width(output_a4)
pdf_height(output_a4)
unlink(output_a4)

unlink(input)
}
