% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pages.R
\name{pdf_pages}
\alias{pdf_pages}
\title{Get integer vector of subset of pdf pages}
\usage{
pdf_pages(input, ..., pages = c("all", "even", "odd"))
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or "all", "even", or "odd" to include all pages, just the even pages,
or just the odd pages.}
}
\value{
An integer vector.
}
\description{
\code{pdf_pages()} calculates an integer vector of subset of pdf pages.
}
\examples{
# Create an 8-page document
f <- tempfile(fileext = ".pdf")
grDevices::pdf(f)
for (page in seq.int(8L)) {
  grid::grid.newpage()
  grid::grid.text(paste("Page", page))
}
invisible(grDevices::dev.off())

pdf_pages(f, pages = 1:4)
pdf_pages(f, pages = -(1:4))
pdf_pages(f, pages = "all")
pdf_pages(f, pages = "even")
pdf_pages(f, pages = "odd")

unlink(f) # Clean up
}
