% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_pages.R
\name{pdf_pages}
\alias{pdf_pages}
\title{Get integer vector of subset of pdf pages}
\usage{
pdf_pages(input, ..., pages = c("all", "even", "odd", "2-up saddle stitch"))
}
\arguments{
\item{input}{Input pdf filename.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}
}
\value{
An integer vector.
}
\description{
\code{pdf_pages()} calculates an integer vector of subset of pdf pages.
}
\examples{
f <- pdf_create_blank(length = 8L)
pdf_pages(f, pages = 1:4)
pdf_pages(f, pages = -(1:4))
pdf_pages(f, pages = "all")
pdf_pages(f, pages = "even")
pdf_pages(f, pages = "odd")

unlink(f) # Clean up
}
