% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_rm_crosshairs.R
\name{pdf_rm_crosshairs}
\alias{pdf_rm_crosshairs}
\title{Remove crosshairs}
\usage{
pdf_rm_crosshairs(
  input,
  output = NULL,
  ...,
  layout = "poker_3x2_bleed",
  pages = "odd",
  dpi = 300
)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{layout}{Either a layout preset name in \code{\link[=layout_names]{layout_names()}} or a data frame
with layout data (as returned by \code{\link[=layout_grid]{layout_grid()}}).}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or a string:\describe{
\item{all}{Include all the pages.}
\item{even}{Include just the even pages.}
\item{odd}{Include just the odd pages.}
\item{2-up saddle stitch}{The order of the pages to create a saddle-stitch booklet if printing 2-up.}
}}

\item{dpi}{Dots per inch.  Passed to \code{\link[pdftools:pdf_render_page]{pdftools::pdf_render_page()}}.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect removes from crosshairs from a pdf.
}
\description{
\code{pdf_rm_crosshairs()} removes unwanted crosshairs.
}
\details{
\itemize{
\item In order to work the PnP layout needs a solid color bleed.
\item The default layout supports \href{https://greggjewell.itch.io/galdors-grip}{Galdor's Grip} (PnP letter size v1).
\item The original pdf document will be rasterized.
}
}
\examples{
if (requireNamespace("bittermelon", quietly = TRUE) &&
    requireNamespace("piecepackr", quietly = TRUE) &&
    utils::packageVersion("piecepackr") >= "1.14.0-5") {
  f1 <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
  f2 <- pdf_add_crosshairs(f1, pages = "all",
                           layout = "poker_3x2_bleed", dpi = 75)
  f3 <- pdf_rm_crosshairs(f2, pages = "odd",
                          layout = "poker_3x2_bleed", dpi = 75)
  unlink(f1)
  unlink(f2)
  unlink(f3)
}
}
