% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_rm_crosshairs.R
\name{pdf_rm_crosshairs}
\alias{pdf_rm_crosshairs}
\title{Remove crosshairs}
\usage{
pdf_rm_crosshairs(input, output = NULL, ..., pages = "odd")
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Ignored.}

\item{pages}{A positive numeric vector of pages to include,
a negative numeric vector of pages to exclude,
or "all", "even", or "odd" to include all pages, just the even pages,
or just the odd pages.}
}
\value{
\code{output} pdf file name invisibly.
As a side effect removes from crosshairs from a pdf.
}
\description{
\code{pdf_rm_crosshairs()} removes unwanted crosshairs.
Currently only supports \href{https://greggjewell.itch.io/galdors-grip}{Galdor's Grip} (PnP letter size v1).
}
\examples{
\donttest{# Make take more than 5s
input <- pdf_create_blank(length = 2L, width = 11, height = 8.5)
output <- pdf_rm_crosshairs(input, pages = "odd")
unlink(input)
unlink(output)
}
}
