% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpmisc-package.R
\docType{package}
\name{pnpmisc}
\alias{pnpmisc-package}
\alias{pnpmisc}
\title{pnpmisc: Utilities for Print-and-Play Board Games}
\description{
Utilities for print-and-play board games.
}
\section{Package options}{

The following \code{pnpmisc} option  may be set globally via \code{\link[base:options]{base::options()}}:
\describe{
\item{pnpmisc.dpi}{Set new default \code{dpi} (dots per inch) used for PDF rasterization.
Defaults to \code{300}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trevorld/pnpmisc}
  \item \url{https://trevorldavis.com/R/pnpmisc/dev/}
  \item Report bugs at \url{https://github.com/trevorld/pnpmisc/issues}
}

}
\author{
\strong{Maintainer}: Trevor L. Davis \email{trevor.l.davis@gmail.com} (\href{https://orcid.org/0000-0001-6341-4639}{ORCID})

}
\keyword{internal}
