% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{pdf_compress}
\alias{pdf_compress}
\alias{pdf_rotate_pages}
\alias{pdf_subset}
\title{Wrappers around qpdf functions}
\usage{
pdf_compress(input, output = NULL, ...)

pdf_rotate_pages(input, output = NULL, ...)

pdf_subset(input, output = NULL, ...)
}
\arguments{
\item{input}{Input pdf filename.}

\item{output}{Output pdf filename.  \code{NULL} defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Passed to the underlying qpdf functions.}
}
\value{
\code{output} filename of new pdf file invisibly.
}
\description{
These functions wrap around the utilities in the qpdf package
but alters the arguments to match this package's conventions:
the first two arguments must be \code{input} and \code{output},
by default \code{output = tempfile(fileext = ".pdf")}, and all other
arguments must be named.
}
\examples{
f1 <- tempfile(fileext = ".pdf")
pdf(f1, width = 6, height = 4)
grid::grid.text("")
invisible(dev.off())

f2 <- pdf_compress(f1)
f3 <- pdf_subset(f2, pages = 1L)
f4 <- pdf_rotate_pages(f3, angle = 90)

unlink(f1)
unlink(f2)
unlink(f3)
unlink(f4)
}
\seealso{
\code{\link[qpdf:qpdf]{qpdf::pdf_compress()}}, \code{\link[qpdf:qpdf]{qpdf::pdf_rotate_pages()}}, \code{\link[qpdf:qpdf]{qpdf::pdf_subset()}}
}
